reshape module structures
