+++ src/endpoint/primitive/result.rs
@@ -1,8 +1,8 @@
 #![allow(missing_docs)]
 
 use std::marker::PhantomData;
+use endpoint::{Endpoint, EndpointContext, EndpointError};
 use task::{self, TaskResult};
-use super::{Endpoint, EndpointContext, EndpointError};
 
 
 pub fn ok<T: Clone, E>(x: T) -> EndpointOk<T, E> {
