+++ src/endpoint/primitive/skip_all.rs
